requirements.txt
git+https://github.com/MycroftKang/highlighter-core.git@fec30ae4b42480a0cc99438db062c8b4871c8a9f
black==20.8b1
drf-yasg[validation]==1.20.0
Django==3.1.5
dj-database-url==0.5.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
pylint==2.6.0
psycopg2-binary==2.8.6
whitenoise==5.2.0
